rename externalbo
